refactor: design issues fix and update readme
