--- Dockerfile.dev
@@ -1,21 +1,0 @@
-FROM node:18-alpine
-
-WORKDIR /app
-
-# Install dependencies
-COPY package.json yarn.lock* package-lock.json* pnpm-lock.yaml* ./
-RUN \
-  if [ -f yarn.lock ]; then yarn install; \
-  elif [ -f package-lock.json ]; then npm install; \
-  elif [ -f pnpm-lock.yaml ]; then corepack enable pnpm && pnpm install; \
-  else echo "Lockfile not found." && exit 1; \
-  fi
-
-# Copy source code
-COPY . .
-
-# Expose port
-EXPOSE 3000
-
-# Start development server
-CMD ["yarn", "dev"]
