+++ app/layout.tsx
@@ -1,6 +1,4 @@
 import type { Metadata, Viewport } from "next";
-import { clsx } from "clsx";
-import { inter, roboto } from "@/configs/font-config";
 import "@/assets/globals.css";
 import { ReactNode } from "react";
 import { GradientBackground } from "@/components/common/gradient-background";
@@ -8,7 +6,7 @@ import localFont from "next/font/local";
 
 const geistSans = localFont({
   src: "./fonts/Geist[wght].woff2",
-  weight: "100 900", // range for variable font
+  weight: "100 900",
   variable: "--font-geist",
   display: "swap",
 });
@@ -19,7 +17,7 @@ export const metadata: Metadata = {
 };
 
 export const viewport: Viewport = {
-  themeColor: "#33a05a",
+  themeColor: "#16191C",
   initialScale: 1,
   maximumScale: 5,
 };
@@ -31,6 +29,9 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
+      <head>
+        <link rel="icon" href="/favicon.svg" type="image/svg+xml" />
+      </head>
       <body className={`${geistSans.variable} font-sans`}>
         <GradientBackground />
         {children}
