+++ components/common/badge.tsx
@@ -1,5 +1,5 @@
 import { twclsx } from "@/utils/twclsx";
-import { FunctionComponent, ReactNode, useState } from "react";
+import { FunctionComponent, ReactNode } from "react";
 
 interface BadgeProps {
   label: string;
@@ -11,7 +11,7 @@ const Badge: FunctionComponent<BadgeProps> = ({ label, icon, className }) => {
   return (
     <span
       className={twclsx(
-        "text-orange-dark bg-orange-light mx-1.5 rounded-3xl px-[3px] text-xs font-medium leading-4 tracking-tight",
+        "mx-1.5 rounded-3xl bg-orange-light px-[3px] text-xs font-medium leading-4 tracking-tight text-orange-dark",
         {
           "gap-1": icon,
         },
