+++ constants/routes.ts
@@ -1,3 +1,2 @@
-// API Routes for MusicGPT Application
 export const API_ROUTE_VOICES = "/api/voices";
 export const API_ROUTE_GENERATE_SONG = "/api/generate-song";
