+++ docker-compose.yml
@@ -1,10 +1,12 @@
 version: "3.8"
 
 services:
+  # Production service
   musicgpt-app:
     build:
       context: .
       dockerfile: Dockerfile
+      target: runner
     ports:
       - "3000:3000"
     environment:
@@ -18,11 +20,12 @@ services:
       retries: 3
       start_period: 40s
 
-  # Development service (optional)
+  # Development service
   musicgpt-dev:
     build:
       context: .
-      dockerfile: Dockerfile.dev
+      dockerfile: Dockerfile
+      target: development
     ports:
       - "3001:3000"
     environment:
